use qmake's c++11 compiler switch macro in .pro
rather than explicitly say -std=c++11.

grepping qt's mkspecs/ directory, i see clang has the macro set to
-std=c++11 and gcc has it set to -std=c++0x. i guess that's because qt
distributes with a mkspec for gcc42, not gcc48. maybe we can overwrite
that, or explain in the readme how to overwrite that when setting up the
project and creating a copy of the mkspec for gcc48.