extract glew, eigen as external libraries in build
 * mv src/VPaint.pro VPaint.pro
 * VPaint.pro is of template subdirs
 * mv src/ app
 * add subproject app/app.pro
 * mv src/Eigen src/GLEW lib/
 * add subproject lib/GLEW/glew.pro
 * change #include "../Eigen/Core" to #include <Eigen/Core>

express external libraries as such in the build, so crucial warnings are
heard.

eigen and glew are not part of the project. they are dependencies,
external libraries. any warnings coming out of there are not interesting
as far as this project is concerned, because only developers of those
projects can fix them. what's worse is warnings coming from external
libraries flood and hide relevant and interesting warnings coming from
code of this project.